clean up, delete unused methods and files, move tree.ts to shared, tree-thems to material
