+++ src/app/admin/admin.component.scss
@@ -1,4 +1,4 @@
-@import '../tree-theme';
+@import '../material/tree-theme';
 
 .admin {
   $table-height: calc(100vh - 360px);
