+++ src/app/app.component.scss
@@ -1,4 +1,4 @@
-@import './tree-theme';
+@import 'material/tree-theme';
 
 html, body {
   width: 100%;
