+++ src/app/app.component.ts
@@ -6,7 +6,6 @@ import { EmailComponent } from './members/email/email.component';
 import { SignupComponent } from './members/signup/signup.component';
 import { AuthService } from './services/auth.service';
 import { MembersComponent } from './members/members.component';
-import { MapClickEvent } from './tree';
 import { AppStateService } from './services/app-state.service';
 
 @Component({
@@ -20,7 +19,6 @@ export class AppComponent implements OnInit {
   container: ViewContainerRef;
   toolComponent: any;
   memberSubComponent = '';
-  mapClickData: MapClickEvent = null;
   cmpRef: ComponentRef<any>;
   loggedIn = false;
 
@@ -31,20 +29,18 @@ export class AppComponent implements OnInit {
   }
 
   ngOnInit() {
-
     this.authService.userChanged.subscribe(
       () => {
         this.loggedIn = this.authService.isLoggedIn();
       }
     );
-
   }
 
   displayInfoPage() {
     document.getElementById('infoPage').style.display = '';
   }
 
-  userHasRole(role: string){
+  userHasRole(role: string) {
     return this.authService.hasRole(role);
   }
 
@@ -52,7 +48,6 @@ export class AppComponent implements OnInit {
   logout() {
     this.authService.logout();
     this.addTool('hide');
-    console.log(this.authService.getUser());
   }
   toggleLogin() {
     if (this.loggedIn) {
@@ -62,27 +57,11 @@ export class AppComponent implements OnInit {
     }
   }
   loginEvent(event) {
-    console.log(event);
     this.addTool(event);
   }
 
   setMode(mode) {
     this.appState.setMode(mode);
-    // console.log('select mode');
-    /* this.memberSubComponent = "mode-selector";
-    this.addTool('member'); */
-  }
-
-  mapClicked(event: MapClickEvent) {
-    console.log(event);
-    this.mapClickData = event;
-    if (event.attr && (this.authService.mode === 'play')) {
-      this.memberSubComponent = "guess";
-      this.addTool('member');
-    } else if (this.authService.mode === 'add') {
-      this.memberSubComponent = "add";
-      this.addTool('member');
-    }
   }
 
   // creates a component and shows it in the browser
@@ -114,12 +93,10 @@ export class AppComponent implements OnInit {
       const comp = this._cfr.resolveComponentFactory(this.toolComponent);
       this.cmpRef = this.container.createComponent(comp);
       this.cmpRef.instance.subComponent = this.memberSubComponent;
-      this.cmpRef.instance.mapClickData = this.mapClickData;
       this.cmpRef.instance.eventData.subscribe(
         (data) => {
           console.log(data);
           this.addTool(data);
-          this.mapClickData = null;
           this.memberSubComponent = '';
         }
       );
