+++ src/app/map/guess-panel/guess-panel.component.scss
@@ -1,4 +1,4 @@
-@import '../../tree-theme';
+@import '../../material/tree-theme';
 
 
 .guess-panel {
