+++ src/app/map/side-panel/side-panel.component.scss
@@ -1,4 +1,4 @@
-@import '../../tree-theme';
+@import '../../material/tree-theme';
 
 .side-panel {
   position: absolute;
