+++ src/app/material/tree-theme.scss
@@ -1,4 +1,4 @@
-@import '../../node_modules/@angular/material/theming';
+@import '../../../node_modules/@angular/material/theming';
 
 @include mat-core();
