+++ src/app/page-not-found/page-not-found.component.scss
@@ -1,4 +1,4 @@
-@import '../tree-theme';
+@import '../material/tree-theme';
 
 .page-not-found {
   font-size: 1.5em;
